+++ packages/typescript/src/components/CallSignature.tsx
@@ -1,9 +1,8 @@
 import { Children } from "@alloy-js/core/jsx-runtime";
 import {
   FunctionDeclaration,
-  TypeParameterDescriptor,
 } from "./FunctionDeclaration.jsx";
-import { ParameterDescriptor } from "../parameter-descriptor.js";
+import { ParameterDescriptor, TypeParameterDescriptor } from "../parameter-descriptor.js";
 
 export interface CallSignatureProps {
   /**
