+++ packages/typescript/src/parameter-descriptor.ts
@@ -15,14 +15,14 @@ export interface ParameterDescriptor {
   type?: Children;
 
   /**
-   * The refkey for this parameter.
+   * The default value of the parameter.
    */
-  refkey?: Refkey;
+  default?: Children;
 
   /**
-   * The refkeys for this parameter.
+   * The refkey for this parameter.
    */
-  refkeys?: Refkey[];
+  refkey?: Refkey | Refkey[];
 
   /**
    * Whether the parameter is optional.
@@ -38,4 +38,36 @@ export interface ParameterDescriptor {
    * Documentation for the parameter.
    */
   doc?: Children;
+}
+
+
+
+/**
+ * Information for a TypeScript generic type parameter.
+ */
+export interface TypeParameterDescriptor {
+  /**
+   * The name of the type parameter.
+   */
+  name: string;
+
+  /**
+   * The extends constraint for the type parameter.
+   */
+  extends?: Children;
+
+  /**
+   * The default type of the type parameter.
+   */
+  default?: Children;
+
+  /**
+   * A refkey or array of refkeys for this type parameter.
+   */
+  refkey?: Refkey | Refkey[];
+
+  /**
+   * Arbitrary metadata for the type parameter symbol.
+   */
+  metadata?: Record<string, unknown>;
 }
